refactor(workflow): split 1522-line workflow.ts into SRP modules
The workflow command class mixed task-lifecycle (now/_showActiveTask),
the rule-management dispatcher (workflow), 12 rule-management
subcommand handlers (_workflowAdd/_workflowGate/_workflowInstruction/
_workflowRm/_workflowReset/_workflowDisable/_workflowHelp/
_workflowShow/_workflowInit/_workflowCreate/_workflowList/
_workflowDelete), intent detection, and 7 file-level helpers.

This commit splits into focused modules under workflow/:
  - workflow/intent.ts        — IntentType, INTENT_PATTERNS, detectIntent,
                                parseAction, searchRules
  - workflow/rule-actions.ts  — 12 standalone subcommand handlers
                                (formerly _workflowX private methods)
  - workflow/md-helpers.ts    — buildFlowDiagram for `--md` output

workflow.ts kept now()/_showActiveTask/workflow(dispatcher)/run() —
went from 1522 → 360 LOC.

Side cleanup: removed 6 file-level helpers that were dead code
(_buildEfficiencySection, _buildRpiSection, _formatMinutesToDuration,
_loadRepoAnalysis, _getFilesModifiedSinceTaskStart, _formatVariance).
Knip's `ignoreExportsUsedInFile` masked them; grep confirms no
callers in or out of the file.

Behavior unchanged:
  - typecheck clean
  - biome clean
  - 1029 tests pass (3465 expects)
  - bun run build succeeds

Follow-up needed: rule-actions.ts has DRY violations (the
`if (options.md) ...; else out.warn(...)` pattern repeats 20+ times,
the "workflow not found" check repeats 3 times, several magic numbers
for timeoutMs). Addressed in a follow-up PR.
